Refactor some package locations
