+++ tasks/realvncserver/rvstbuilder/rvstbuilder.go
@@ -1,9 +1,9 @@
 package rvstbuilder
 
 import (
-	builder "github.com/realvnc-labs/tacoscript/builder"
 	"github.com/realvnc-labs/tacoscript/tasks"
 	"github.com/realvnc-labs/tacoscript/tasks/realvncserver"
+	"github.com/realvnc-labs/tacoscript/tasks/shared/builder"
 	"github.com/realvnc-labs/tacoscript/tasks/shared/fieldstatus"
 )
