+++ tasks/support/pkgmanager/diff_test.go
