+++ tasks/support/pkgmanager/manager_test.go
