+++ tasks/support/reg/deletekey_win.go
