+++ tasks/support/reg/registry_common.go
