+++ tasks/support/reg/registry_other.go
