+++ tasks/support/reg/registry_test.go
