+++ tasks/support/reg/registry_win.go
