+++ tasks/winreg/wrtask.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	tacoexec "github.com/realvnc-labs/tacoscript/exec"
-	"github.com/realvnc-labs/tacoscript/reg"
 	"github.com/realvnc-labs/tacoscript/tasks"
 	"github.com/realvnc-labs/tacoscript/tasks/shared/executionresult"
+	"github.com/realvnc-labs/tacoscript/tasks/support/reg"
 
 	"github.com/realvnc-labs/tacoscript/utils"
